PR Feedback
Change flag names and document default values
Remove callWithTelemetry
Move callWithTelemetryAndErrorHandling to separate file
