+++ ui/src/BaseEditor.ts
@@ -59,10 +59,10 @@ export abstract class BaseEditor<ContextT> implements vscode.Disposable {
     }
 
     public async onDidSaveTextDocument(actionContext: IActionContext, globalState: vscode.Memento, doc: vscode.TextDocument): Promise<void> {
-        actionContext.sendTelemetry = false;
+        actionContext.suppressTelemetry = true;
         const filePath: string | undefined = Object.keys(this.fileMap).find((fsPath: string) => path.relative(doc.uri.fsPath, fsPath) === '');
         if (!this.ignoreSave && filePath) {
-            actionContext.sendTelemetry = true;
+            actionContext.suppressTelemetry = false;
             const context: ContextT = this.fileMap[filePath][1];
             const showSaveWarning: boolean | undefined = vscode.workspace.getConfiguration().get(this.showSavePromptKey);
