+++ ui/src/index.ts
@@ -8,3 +8,4 @@ export { UserCancelledError } from './errors';
 export { BaseEditor } from './BaseEditor';
 export { AzureActionHandler } from './AzureActionHandler';
 export { parseError } from './parseError';
+export { callWithTelemetryAndErrorHandling } from './callWithTelemetryAndErrorHandling';
